+++ hil-test/tests/delay.rs
@@ -7,7 +7,10 @@
 #![no_main]
 
 use embedded_hal::delay::DelayNs;
-use esp_hal::delay::Delay;
+use esp_hal::{
+    delay::Delay,
+    time::{Duration, Instant},
+};
 use hil_test as _;
 
 esp_bootloader_esp_idf::esp_app_desc!();
@@ -19,6 +22,7 @@ struct Context {
 #[cfg(test)]
 #[embedded_test::tests(default_timeout = 2)]
 mod tests {
+
     use super::*;
 
     #[init]
@@ -30,10 +34,17 @@ mod tests {
     }
 
     #[test]
+    fn duration_since_epoch_is_not_relative_to_now(mut ctx: Context) {
+        let now = Instant::EPOCH;
+        ctx.delay.delay_ns(10_000);
+        assert_eq!(now.duration_since_epoch(), Duration::ZERO);
+    }
+
+    #[test]
     fn delay_ns(mut ctx: Context) {
-        let t1 = esp_hal::time::Instant::now();
+        let t1 = Instant::now();
         ctx.delay.delay_ns(600_000);
-        let t2 = esp_hal::time::Instant::now();
+        let t2 = Instant::now();
 
         assert!(t2 > t1);
         assert!(
@@ -45,9 +56,9 @@ mod tests {
 
     #[test]
     fn delay_70millis(ctx: Context) {
-        let t1 = esp_hal::time::Instant::now();
+        let t1 = Instant::now();
         ctx.delay.delay_millis(70);
-        let t2 = esp_hal::time::Instant::now();
+        let t2 = Instant::now();
 
         assert!(t2 > t1);
         assert!(
@@ -59,9 +70,9 @@ mod tests {
 
     #[test]
     fn delay_1_500us(mut ctx: Context) {
-        let t1 = esp_hal::time::Instant::now();
+        let t1 = Instant::now();
         ctx.delay.delay_us(1_500);
-        let t2 = esp_hal::time::Instant::now();
+        let t2 = Instant::now();
 
         assert!(t2 > t1);
         assert!(
@@ -73,9 +84,9 @@ mod tests {
 
     #[test]
     fn delay_3_00ms(mut ctx: Context) {
-        let t1 = esp_hal::time::Instant::now();
+        let t1 = Instant::now();
         ctx.delay.delay_ms(300);
-        let t2 = esp_hal::time::Instant::now();
+        let t2 = Instant::now();
 
         assert!(t2 > t1);
         assert!(